backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
requests==2.31.0
xgboost==2.0.3
scikit-learn==1.4.0
joblib==1.3.2
numpy==1.26.3
httpx==0.26.0
python-dotenv==1.0.0
google-generativeai>=0.3.0
edge-tts>=7.0.0
